Create new API for testing, see #6925
