+++ lib/sidekiq.rb
@@ -47,6 +47,14 @@ module Sidekiq
     puts "Take a deep breath and count to ten..."
   end
 
+  def self.testing!(mode = :fake, &block)
+    raise "Sidekiq.testing not allowed in production" if Sidekiq.default_configuration[:environment] == "production"
+    raise "Unknown testing mode: #{mode}" unless %i[fake disabled inline].include?(mode)
+
+    require "sidekiq/test_api"
+    Sidekiq::Testing.__set_test_mode(mode, &block)
+  end
+
   def self.server?
     defined?(Sidekiq::CLI)
   end
